refactor: move SpinalSimOneEntryRamTester
